refactor: move remaining AssetLoader functions into resourceLoaderFunctions.ts
