+++ packages/engine/src/assets/AssetModule.ts
@@ -23,7 +23,6 @@ All portions of the code written by the Infinite Reality Engine team are Copyrig
 Infinite Reality Engine. All Rights Reserved.
 */
 
-import { AssetLoader } from './classes/AssetLoader'
 import { AssetLoaderState } from './state/AssetLoaderState'
 import { DomainConfigState } from './state/DomainConfigState'
 import { ResourceLoadingManagerState } from './state/ResourceLoadingManagerState'
@@ -31,6 +30,5 @@ import { ResourceLoadingManagerState } from './state/ResourceLoadingManagerState
 export default {
   AssetLoaderState,
   DomainConfigState,
-  ResourceLoadingManagerState,
-  AssetLoader
+  ResourceLoadingManagerState
 }
