+++ packages/engine/src/scene/components/UVOL2Component.ts
@@ -70,7 +70,7 @@ import { isMobileXRHeadset } from '@ir-engine/spatial/src/xr/XRState'
 import { S } from '@ir-engine/ecs/src/schemas/JSONSchemas'
 import { AssetExt } from '@ir-engine/engine/src/assets/constants/AssetType'
 import { MeshComponent } from '@ir-engine/spatial/src/renderer/components/MeshComponent'
-import { getLoader } from '../../assets/classes/AssetLoader'
+import { getLoader } from '../../assets/functions/resourceLoaderFunctions'
 import { AssetLoaderState } from '../../assets/state/AssetLoaderState'
 import { AudioState } from '../../audio/AudioState'
 import { AssetState } from '../../gltf/GLTFState'
